Change file structure to suite node server
In light of the use of a node web server the file structure
had to be changed to match what is hosted on the server.
An app folder was made for the files that would be sent to the client
from the web server. All other project files outside the app folder
is not neccesarry for the client.

Other Changes

- Pages/event.html -> app/index.html

This was neccesarry so that the webserver could immeditely load
a webpage instead of a file tree.

Resolves: #3
See also: #2